add dependencies of music bot

+++ cmds/music.py
@@ -6,17 +6,22 @@ from core import Cog_Extension
 class Music(Cog_Extension):
         
     @commands.command()
-    async def play(self, ctx, url):
+    async def play(self, ctx, *args):
+        if len(args) != 1:
+            await ctx.send("不合法的用法: 用法: play <連結>")
+            return
+        
+        url = args[0]
         song_exist = os.path.isfile("song.mp3")
         try:
             if song_exist:
                 os.remove("song.mp3")
         except PermissionError:
-            await ctx.send("請等待目前歌曲結束 ，或者使用 'stop' 指令")
+            await ctx.send("請等待目前歌曲結束，或者使用 'stop' 指令")
             return
         
         '''
-        TODO (optional)
+        TODO (bonus)
         讓使用者能夠查詢歌曲
         (可以使用爬蟲，或者參考 yt-dlp 的更多功能)
         Reference : https://www.mankier.com/1/yt-dlp
@@ -27,31 +32,31 @@ class Music(Cog_Extension):
 
         if voice is None:
             voiceChannel = discord.utils.get(ctx.guild.voice_channels, name='General')
-            await voiceChannel.connect(timeout = 600.0)
+            await voiceChannel.connect(timeout=600.0)
             voice = discord.utils.get(self.bot.voice_clients, guild=ctx.guild)
 
         for file in os.listdir("./"):
             if file.endswith(".mp3"):
                 os.rename(file, "song.mp3")
     
-        voice.play(discord.FFmpegPCMAudio(executable = 'ffmpeg.exe', source = "song.mp3"))
+        voice.play(discord.FFmpegPCMAudio(executable='ffmpeg.exe', source="song.mp3"))
 
     '''
-        TODO (optional)
-        新增其他功能，如歌曲佇列、投票等等
+    TODO (bonus)
+    讓使用者管理待播放清單，包含加入歌曲、插歌...
     '''
     
     @commands.command()
-    async def leave(self,ctx):
+    async def leave(self, ctx):
         voice = discord.utils.get(self.bot.voice_clients, guild=ctx.guild)
         try:
-                await voice.disconnect()
+            await voice.disconnect()
         except:
             await ctx.send("機器人沒有連接到語音頻道")
 
 
     @commands.command()
-    async def pause(self,ctx):
+    async def pause(self, ctx):
         voice = discord.utils.get(self.bot.voice_clients, guild=ctx.guild)
         try: 
             if voice.is_playing():
@@ -63,7 +68,7 @@ class Music(Cog_Extension):
 
 
     @commands.command()
-    async def resume(self,ctx):
+    async def resume(self, ctx):
         voice = discord.utils.get(self.bot.voice_clients, guild=ctx.guild)
         try:
             if voice.is_paused():
@@ -74,7 +79,7 @@ class Music(Cog_Extension):
             await ctx.send("機器人沒有連接到語音頻道")
 
     @commands.command()
-    async def stop(self,ctx):
+    async def stop(self, ctx):
         voice = discord.utils.get(self.bot.voice_clients, guild=ctx.guild)
         try:
             voice.stop()
